backend/requirements.txt
cprint==1.2.2
dataclasses==0.6
dataclasses-json==0.5.7
requests==2.29.0
fastapi==0.95.2
python-dotenv==1.0.0
hypercorn==0.14.3